+++ CHANGELOG.md
@@ -13,6 +13,8 @@ was breaking things when these were run with a stopped docker service.
 - Resolved a potential failure case where the host mount path is not reported
 as a native path. That would prevent proper container discovery. This was not
 being an issue previously but I realized it was not working for Windows now.
+- Do not throw if the output stream is not finalized. This was previously causing
+issues when the long output of a command is piped to another process like `less`.
 
 ### Changed